fix(settings): resolve integration review and CI regressions
