+++ desktop/frontend/src/__tests__/provider-editor-model-picker.test.tsx
@@ -317,7 +317,7 @@ const providerUrlLabel = Array.from(rootEl.querySelectorAll<HTMLLabelElement>("l
   (label) => label.htmlFor === providerUrlInput?.id,
 );
 ok(Boolean(providerUrlLabel) && providerUrlInput?.getAttribute("aria-describedby") !== null, "provider URL input has a programmatic label and description");
-ok(rootEl.textContent?.includes("Request URL:") === true, "provider URL helper explains exact request behavior");
+ok(rootEl.textContent?.includes("uses it unchanged") === true, "provider URL helper explains exact request behavior");
 ok(rootEl.querySelector<HTMLInputElement>('input[placeholder="e.g. my-proxy"]')?.disabled !== true, "new custom provider name stays editable");
 ok(rootEl.querySelector<HTMLInputElement>('input[placeholder="e.g. my-proxy"]')?.nextElementSibling?.classList.contains("mem-hint") !== true, "new custom provider editor omits the rename hint");
 
@@ -351,7 +351,7 @@ ok(
   "backend text-only metadata omits image badges outside the legacy endpoint heuristic",
 );
 const displayedRequestURL = (input: HTMLInputElement | null) =>
-  document.getElementById(input?.getAttribute("aria-describedby") ?? "")?.querySelector("code")?.textContent;
+  input?.value;
 const customProviderUrlInput = rootEl.querySelector<HTMLInputElement>(".provider-url-input");
 ok(rootEl.querySelectorAll('input[type="radio"]').length === 0, "existing custom providers no longer expose an address mode selector");
 ok(customProviderUrlInput?.value === "https://eu.deepseek.com/v1/chat/completions", "legacy base-only providers display their previously effective request URL");
@@ -366,7 +366,7 @@ await act(async () => {
   await flushPromises();
 });
 const legacyProviderUrlInput = rootEl.querySelector<HTMLInputElement>(".provider-url-input");
-ok(legacyProviderUrlInput?.value === legacyChatURLProvider.baseUrl && displayedRequestURL(legacyProviderUrlInput) === "https://legacy.example.com/chat/completions", "legacy OpenAI chat URLs display their historically normalized effective endpoint");
+ok( displayedRequestURL(legacyProviderUrlInput) === "https://legacy.example.com/chat/completions", "legacy OpenAI chat URLs display their historically normalized effective endpoint");
 const saveButton = Array.from(rootEl.querySelectorAll<HTMLButtonElement>("button")).find(
   (button) => button.textContent?.trim() === "Save changes",
 );
@@ -404,10 +404,11 @@ await act(async () => {
   root.render(renderProviderEditor({ ...legacyChatURLProvider, name: "save-failure" }, () => { throw new Error("storage unavailable"); }));
   await flushPromises();
 });
-const failedSaveButton = Array.from(rootEl.querySelectorAll<HTMLButtonElement>("button")).find((button) => button.textContent?.trim() === "Save");
+const failedSaveButton = Array.from(rootEl.querySelectorAll<HTMLButtonElement>("button")).find((button) => button.textContent?.trim() === "Save changes");
+await act(async () => editDisplayName());
 await act(async () => { failedSaveButton?.click(); await flushPromises(); });
 ok(rootEl.textContent?.includes("storage unavailable") === true, "save errors remain visible in the provider form");
-ok(rootEl.querySelector<HTMLInputElement>(".provider-url-input")?.value === legacyChatURLProvider.baseUrl && rootEl.querySelector(".provider-model-row") !== null, "failed save retains connection and model drafts");
+ok(rootEl.querySelector<HTMLInputElement>(".provider-url-input")?.value === "https://legacy.example.com/chat/completions" && rootEl.querySelector(".provider-model-draft__option") !== null, "failed save retains connection and model drafts");
 
 await act(async () => {
   root.unmount();
