+++ desktop/frontend/src/lib/types.ts
@@ -1,3 +1,5 @@
+import type { ProviderPresetView } from "./providerCatalogTypes";
+export type { ProviderProtocolEndpoint, ProviderCatalog, ProviderPresetView } from "./providerCatalogTypes";
 import type { RecoveryEventFields } from "./recoveryStatus";
 // Wire contract — mirrors desktop/wire.go (itself mirroring internal/serve/wire.go).
 // One event channel carries every kind; `kind` discriminates the payload.
@@ -1804,50 +1806,7 @@ export interface ProviderModelCapabilityUpdate {
   inputModalities: string[];
 }
 
-export interface ProviderProtocolEndpoint {
-  baseUrl: string;
-  source: string;
-  checkedOn: string;
-  authHeader?: boolean;
-  responsesMode?: string;
-}
 
-export interface ProviderCatalog {
-  protocols?: Record<string, ProviderProtocolEndpoint>;
-  brandId: string;
-  brandLabel: string;
-  region: string;
-  product: string;
-  format: string;
-  baseUrl: string;
-}
-
-export interface ProviderPresetView {
-  catalog?: ProviderCatalog;
-  id: string;
-  label: string;
-  description: string;
-  keyEnv: string;
-  recommended?: boolean;
-  billingMode?: string;
-  displayGroup?: string;
-  displaySection?: string;
-  displayTier?: "primary" | "advanced" | "compatibility" | string;
-  routeKind?: string;
-  optional?: boolean;
-  displayOrder?: number;
-  providerNames: string[];
-  models: string[];
-  added: boolean;
-  status?: "available" | "installed" | "installed_modified" | "partial" | "name_conflict" | "similar_existing";
-  statusProviderNames?: string[];
-  missingProviderNames?: string[];
-  keySet: boolean;
-  requiresKey?: boolean;
-  configured?: boolean;
-  keySource?: string;
-  keySourcePath?: string;
-}
 
 export interface ProviderModelOverrideView {
   model: string;
