+++ desktop/frontend/src/locales/en.ts
@@ -2889,14 +2889,10 @@ export const en = {
   "settings.connections.restartRequired": "The desktop backend has not loaded the rename API. Stop it with Ctrl+C in the launch terminal, then run wails dev again and retry.",
   "settings.connections.rename": "Rename connection",
   "settings.connections.renameFailed": "Could not save. Please retry.",
-  "settings.models.check": "Check model endpoint",
-  "settings.models.checked": "Model discovery is reachable; inference calls have not been tested.",
   "settings.models.inherit": "Inherited",
   "settings.models.saveKeyFirst": "Save the connection and key before checking or fetching models. This action does not save keys.",
-  "settings.models.changeKey": "Change",
   "settings.models.keySaveFailed": "Could not save the key. Please retry.",
   "settings.models.add": "Add model",
-  "settings.models.ids": "Model IDs (comma-separated)",
   "settings.modelDialog.image": "Image",
   "settings.modelDialog.video": "Video",
   "settings.modelDialog.unavailable": "Direct video and PDF input is not available yet.",
@@ -2910,8 +2906,6 @@ export const en = {
   "settings.modelDialog.resetOutput": "Reset output limit",
   "settings.modelDialog.outputLimit": "Output limit (tokens)",
   "settings.modelDialog.outputHint": "Blank inherits; -1 omits the optional request limit. Model limits still apply.",
-  "settings.modelDialog.override": "Override image capability",
-  "settings.modelDialog.auto": "Use detected capability",
   "settings.modelDialog.overrideHint": "Overrides the capability declaration, not what the model can actually do.",
   "settings.modelDialog.capabilities": "Model capabilities",
   "settings.modelDialog.input": "Input",
@@ -2923,15 +2917,9 @@ export const en = {
   "settings.modelDialog.apply": "Apply changes",
   "settings.modelDialog.add": "Add to list",
   "settings.models.saveChanges": "Save changes",
-  "settings.models.connectionConfig": "Connection settings",
   "settings.models.unsaved": "Unsaved changes",
   "settings.models.saved": "No unsaved changes",
-  "settings.models.contextHint": "Leave blank to inherit. Changes take effect when you save.",
-  "settings.models.draftHint": "Checked models are enabled. Changes take effect after saving configuration.",
-  "settings.models.fetchHint": "Fetch candidates while retaining selections. Add models manually if discovery is unsupported.",
-  "settings.models.keyActions": "Key actions",
   "settings.connections.copy": "Copy connection without key",
-  "settings.models.keyHint": "Saved separately for this connection only.",
   "settings.models.options": "Model settings",
   "settings.connections.name": "Connection name",
   "settings.connections.id": "Internal ID (unchanged)",
