+++ desktop/frontend/src/locales/zh-TW.ts
@@ -2014,14 +2014,10 @@ export const zhTW: Record<DictKey, string> = {
   "settings.connections.restartRequired": "目前桌面後端尚未載入重新命名介面。請在啟動終端按 Ctrl+C 停止，再執行 wails dev 後重試。",
   "settings.connections.rename": "重新命名連線",
   "settings.connections.renameFailed": "儲存失敗，請重試。",
-  "settings.models.check": "檢查模型介面",
-  "settings.models.checked": "模型清單介面可存取；未驗證模型推理呼叫。",
   "settings.models.inherit": "繼承",
   "settings.models.saveKeyFirst": "請先儲存連線與金鑰，再檢查或取得模型；此操作不會自動儲存金鑰。",
-  "settings.models.changeKey": "更換",
   "settings.models.keySaveFailed": "金鑰儲存失敗，請重試。",
   "settings.models.add": "新增模型",
-  "settings.models.ids": "模型 ID（多個以逗號分隔）",
   "settings.modelDialog.image": "圖片",
   "settings.modelDialog.video": "影片",
   "settings.modelDialog.unavailable": "暫不支援直接輸入影片或 PDF。",
@@ -2035,8 +2031,6 @@ export const zhTW: Record<DictKey, string> = {
   "settings.modelDialog.resetOutput": "重設輸出上限",
   "settings.modelDialog.outputLimit": "輸出上限（Token）",
   "settings.modelDialog.outputHint": "留空繼承；-1 不傳送可選輸出限制，仍受模型實際限制。",
-  "settings.modelDialog.override": "覆寫圖片能力",
-  "settings.modelDialog.auto": "使用自動識別",
   "settings.modelDialog.overrideHint": "僅覆寫能力宣告，不會改變模型實際能力。",
   "settings.modelDialog.capabilities": "模型能力",
   "settings.modelDialog.input": "輸入",
@@ -2048,15 +2042,9 @@ export const zhTW: Record<DictKey, string> = {
   "settings.modelDialog.apply": "套用變更",
   "settings.modelDialog.add": "新增至列表",
   "settings.models.saveChanges": "儲存變更",
-  "settings.models.connectionConfig": "連線設定",
   "settings.models.unsaved": "未儲存變更",
   "settings.models.saved": "無未儲存變更",
-  "settings.models.contextHint": "留空使用預設值；儲存設定後生效。",
-  "settings.models.draftHint": "勾選表示啟用；修改將在儲存設定後生效。",
-  "settings.models.fetchHint": "取得候選模型並保留選擇；不支援時可手動新增。",
-  "settings.models.keyActions": "金鑰操作",
   "settings.connections.copy": "複製連線（不複製金鑰）",
-  "settings.models.keyHint": "金鑰單獨儲存，僅影響目前連線。",
   "settings.models.options": "模型設定",
   "settings.connections.name": "連線名稱",
   "settings.connections.id": "內部識別碼（保持不變）",
