+++ desktop/frontend/src/locales/zh.ts
@@ -2892,14 +2892,10 @@ export const zh: Record<DictKey, string> = {
   "settings.connections.restartRequired": "当前桌面后端尚未加载重命名接口。请在启动终端按 Ctrl+C 停止，再运行 wails dev 后重试。",
   "settings.connections.rename": "重命名连接",
   "settings.connections.renameFailed": "保存失败，请重试。",
-  "settings.models.check": "检查模型接口",
-  "settings.models.checked": "模型列表接口可访问；未验证模型推理调用。",
   "settings.models.inherit": "继承",
   "settings.models.saveKeyFirst": "请先保存连接与密钥，再检查或获取模型；此操作不会自动保存密钥。",
-  "settings.models.changeKey": "更换",
   "settings.models.keySaveFailed": "密钥保存失败，请重试。",
   "settings.models.add": "添加模型",
-  "settings.models.ids": "模型 ID（多个用逗号分隔）",
   "settings.modelDialog.image": "图片",
   "settings.modelDialog.video": "视频",
   "settings.modelDialog.unavailable": "暂不支持直接输入视频或 PDF。",
@@ -2913,8 +2909,6 @@ export const zh: Record<DictKey, string> = {
   "settings.modelDialog.resetOutput": "重置输出上限",
   "settings.modelDialog.outputLimit": "输出上限（Token）",
   "settings.modelDialog.outputHint": "留空继承；-1 不发送可选输出限制，仍受模型实际限制。",
-  "settings.modelDialog.override": "覆盖图片能力",
-  "settings.modelDialog.auto": "使用自动识别",
   "settings.modelDialog.overrideHint": "仅覆盖能力声明，不会改变模型实际能力。",
   "settings.modelDialog.capabilities": "模型能力",
   "settings.modelDialog.input": "输入",
@@ -2926,15 +2920,9 @@ export const zh: Record<DictKey, string> = {
   "settings.modelDialog.apply": "应用更改",
   "settings.modelDialog.add": "添加到列表",
   "settings.models.saveChanges": "保存更改",
-  "settings.models.connectionConfig": "连接配置",
   "settings.models.unsaved": "未保存更改",
   "settings.models.saved": "无未保存更改",
-  "settings.models.contextHint": "留空使用默认值；保存配置后生效。",
-  "settings.models.draftHint": "勾选表示启用；修改将在保存配置后生效。",
-  "settings.models.fetchHint": "获取候选模型，保留已有选择；不支持获取时可手动添加。",
-  "settings.models.keyActions": "密钥操作",
   "settings.connections.copy": "复制连接（不复制密钥）",
-  "settings.models.keyHint": "密钥单独保存，仅影响当前连接。",
   "settings.models.options": "模型设置",
   "settings.connections.name": "连接名称",
   "settings.connections.id": "内部标识（保持不变）",
