+++ docs/REASONING_CONTRACT.zh-CN.md
@@ -13,8 +13,9 @@
 
 `auto` 保留原有“清除覆盖、使用默认”的界面与 CLI 含义，不等于自适应思考。
 请求级 override 使用空字符串表示继承，而不是发送字面值 `auto`。保留旧配置加载时
-对大小写及已退役 `off` 的兼容处理；已有合法 ID 和 TOML 字段名不变。旧的不支持
-别名现在会明确报错，不再静默改档。非法默认值保留供校验，不再替换为第一档。
+对大小写及已退役 `off` 的兼容处理；已有合法 ID 和 TOML 字段名不变。未声明自定义
+档位时，已保存的 DeepSeek `medium`、`xhigh` 沿用历史请求值 `high`，不重写配置。
+新的显式选择和请求覆盖仍拒绝未声明的别名。其他非法值明确报错；非法默认值保留供校验，不替换为第一档。
 
 | 边界 | 兼容行为 |
 | --- | --- |
