+++ internal/config/render.go
@@ -319,10 +319,7 @@ func RenderTOMLForScope(c *Config, scope RenderScope) string {
 			if p.ModelsURL != "" {
 				fmt.Fprintf(&b, "models_url  = %q   # auto-fetch models from this URL on startup\n", p.ModelsURL)
 			}
-			fmt.Fprintf(&b, "api_key_env = %q\n", p.APIKeyEnv)
-			if p.DisplayName != "" {
-				fmt.Fprintf(&b, "display_name = %q\n", p.DisplayName)
-			}
+			renderProviderIdentity(&b, p.APIKeyEnv, p.DisplayName)
 			if p.PresetID != "" {
 				fmt.Fprintf(&b, "preset_id   = %q   # curated preset identity; settings UI uses it to avoid duplicate installs\n", p.PresetID)
 			}
@@ -1002,10 +999,7 @@ func RenderTOMLProjectDelta(c *Config) string {
 			if p.ModelsURL != "" {
 				fmt.Fprintf(&b, "models_url  = %q\n", p.ModelsURL)
 			}
-			fmt.Fprintf(&b, "api_key_env = %q\n", p.APIKeyEnv)
-			if p.DisplayName != "" {
-				fmt.Fprintf(&b, "display_name = %q\n", p.DisplayName)
-			}
+			renderProviderIdentity(&b, p.APIKeyEnv, p.DisplayName)
 			if p.PresetID != "" {
 				fmt.Fprintf(&b, "preset_id   = %q\n", p.PresetID)
 			}
