+++ internal/provider/responses/responses.go
@@ -36,33 +36,6 @@ func init() {
 	provider.Register("dashscope-responses", newFromConfig)
 }
 
-func newFromConfig(cfg provider.Config) (provider.Provider, error) {
-	effort, _ := cfg.Extra["effort"].(string)
-	mode, _ := cfg.Extra["mode"].(string)
-	webSearch, _ := cfg.Extra["web_search"].(bool)
-	var stateful *bool
-	switch value := cfg.Extra["stateful"].(type) {
-	case bool:
-		stateful = &value
-	case *bool:
-		stateful = value
-	}
-	proxy, _ := cfg.Extra["proxy_spec"].(netclient.ProxySpec)
-	keyEnv, _ := cfg.Extra["api_key_env"].(string)
-	keySource, _ := cfg.Extra["api_key_source"].(string)
-	maxOutputTokens, _ := cfg.Extra["max_output_tokens"].(int)
-	requestURL, _ := cfg.Extra["request_url"].(string)
-	return New(Config{
-		Name: cfg.Name, APIKey: cfg.APIKey, BaseURL: cfg.BaseURL, Model: cfg.Model,
-		ModelInfo: cfg.ModelInfo,
-		Effort:    effort, Mode: mode, Stateful: stateful, WebSearch: webSearch, Proxy: proxy,
-		KeyEnv: keyEnv, KeySource: keySource, MaxOutputTokens: maxOutputTokens, RequestURL: requestURL,
-		// Extra 原样透传：vision 等能力开关由调用方（boot/CLI）写入
-		// cfg.Extra，factory 若丢弃则 New() 读不到（评审 #7234 第 3 点）。
-		Extra: cfg.Extra,
-	}), nil
-}
-
 // Config holds Responses API provider settings.
 type Config struct {
 	Name       string
